2c3s9d-python-testcontainers-integration-test-suite-for-payment-microservice/repository_after/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
asyncpg>=0.29.0
aioredis>=2.0.1
aio-pika>=9.3.0
httpx>=0.25.0
python-multipart>=0.0.6
